Relocate ssh secret tools
